requirements.txt
flask
flask-cors
requests
python-dotenv
sqlalchemy
beautifulsoup4 requests
